requirements.txt
urllib3==1.26.11
beautifulsoup4==4.11.2
lxml==4.9.2
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-jet-reboot==1.3.1
djongo==1.3.6
dnspython==2.2.1
idna==3.3
Pillow==9.2.0
pymongo==3.12.3
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.11
uuid==1.30
psycopg2-binary
python-dotenv
whitenoise
gunicorn
django-import_export
pandas
